backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-google-genai==1.0.1
langchain-community==0.0.10
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
reportlab==4.0.7
python-multipart==0.0.6
pydantic==2.5.0
lxml==4.9.3
python-dotenv==1.0.0
